Revert "change bucket policy schema to fit AWS json"
This reverts commit 5cdbbe16b92868b8058cdac7d4208feaf1783369.
