+++ src/test/system_tests/ceph_s3_tests/s3-tests-lists/s3_tests_pending_list.txt
@@ -41,6 +41,7 @@ s3tests_boto3/functional/test_s3.py::test_lifecycle_transition_set_invalid_date
 s3tests_boto3/functional/test_s3.py::test_put_obj_enc_conflict_c_s3
 s3tests_boto3/functional/test_s3.py::test_put_obj_enc_conflict_c_kms
 s3tests_boto3/functional/test_s3.py::test_put_obj_enc_conflict_s3_kms
+s3tests_boto3/functional/test_s3.py::test_bucket_policy_put_obj_s3_noenc
 s3tests_boto3/functional/test_s3.py::test_bucket_policy_put_obj_s3_kms
 s3tests_boto3/functional/test_s3.py::test_bucket_policy_put_obj_kms_noenc
 s3tests_boto3/functional/test_s3.py::test_bucket_policy_put_obj_kms_s3
