+++ src/upgrade/upgrade_scripts/5.14.0/upgrade_bucket_policy.js
@@ -2,7 +2,7 @@
 "use strict";
 
 const util = require('util');
-const { OP_NAME_TO_ACTION } = require('../../../endpoint/s3/s3_bucket_policy_utils');
+const { OP_NAME_TO_ACTION } = require('../../../endpoint/s3/s3_utils');
 
 function _create_actions_map() {
     const actions_map = new Map();
